agent/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.11
djangorestframework==3.15.1
environs==11.0.0
gunicorn==21.2.0
idna==3.6
marshmallow==3.21.1
packaging==24.0
python-dotenv==1.0.1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.10.0
urllib3==2.2.1
